Setup devx-indexers-and-rpcs team in CODEOWNERS (#22666)
## Description 

This adds the newly created `devx-indexers-and-rpcs` team to the crates
the team owns in CODEOWNERS.

## Test plan 

The `devx-indexers-and-rpcs` team needs permissions set up before this
is merged. The permissions can be verified by being able to add the team
as a reviewer to this PR.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: